Move Parser test to ComponentParser.test.js
